resolveai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
loguru==0.7.2
requests==2.32.3
openai==1.24.0        # NEW
python-dotenv==1.0.1  # NEW  (optional convenience for local dev)